@@ -1,27 +1,27 @@
 {
   "name": "laravel/laravel",
   "type": "project",
-  "description": "The Laravel Framework.",
+  "description": "The skeleton application for the Laravel framework.",
   "keywords": [
-    "framework",
-    "laravel"
+    "laravel",
+    "framework"
   ],
   "license": "MIT",
   "require": {
-    "php": "^8.0.2",
-    "fruitcake/laravel-cors": "^2.0.5",
+    "php": "^8.1",
     "guzzlehttp/guzzle": "^7.2",
-    "laravel/framework": "^9.0",
-    "laravel/sanctum": "^2.14",
-    "laravel/tinker": "^2.7"
+    "laravel/framework": "^10.23",
+    "laravel/sanctum": "^3.2",
+    "laravel/tinker": "^2.8"
   },
   "require-dev": {
     "fakerphp/faker": "^1.9.1",
-    "laravel/sail": "^1.0.1",
+    "laravel/pint": "^1.0",
+    "laravel/sail": "^1.25",
     "mockery/mockery": "^1.4.4",
-    "nunomaduro/collision": "^6.1",
-    "phpunit/phpunit": "^9.5.10",
-    "spatie/laravel-ignition": "^1.0"
+    "nunomaduro/collision": "^7.0",
+    "phpunit/phpunit": "^10.1",
+    "spatie/laravel-ignition": "^2.0"
   },
   "autoload": {
     "psr-4": {
@@ -58,7 +58,11 @@
   "config": {
     "optimize-autoloader": true,
     "preferred-install": "dist",
-    "sort-packages": true
+    "sort-packages": true,
+    "allow-plugins": {
+      "pestphp/pest-plugin": true,
+      "php-http/discovery": true
+    }
   },
   "minimum-stability": "dev",
   "prefer-stable": true
