@@ -8,20 +8,16 @@ class AppServiceProvider extends ServiceProvider
 {
   /**
    * Register any application services.
-   *
-   * @return void
    */
-  public function register()
+  public function register(): void
   {
     //
   }
 
   /**
    * Bootstrap any application services.
-   *
-   * @return void
    */
-  public function boot()
+  public function boot(): void
   {
     //
   }
