@@ -8,20 +8,16 @@ class MenuServiceProvider extends ServiceProvider
 {
   /**
    * Register services.
-   *
-   * @return void
    */
-  public function register()
+  public function register(): void
   {
     //
   }
 
   /**
    * Bootstrap services.
-   *
-   * @return void
    */
-  public function boot()
+  public function boot(): void
   {
     $verticalMenuJson = file_get_contents(base_path('resources/menu/verticalMenu.json'));
     $verticalMenuData = json_decode($verticalMenuJson);
