@@ -18,8 +18,8 @@
           <!-- Logo -->
           <div class="app-brand justify-content-center">
             <a href="{{url('/')}}" class="app-brand-link gap-2">
-              <span class="app-brand-logo demo">@include('_partials.macros',['width'=>25,'withbg' => "#696cff"])</span>
-              <span class="app-brand-text demo text-body fw-bolder">{{ config('variables.templateName') }}</span>
+              <span class="app-brand-logo demo">@include('_partials.macros',["width"=>25,"withbg"=>'var(--bs-primary)'])</span>
+              <span class="app-brand-text demo text-body fw-bold">{{ config('variables.templateName') }}</span>
             </a>
           </div>
           <!-- /Logo -->
