@@ -15,7 +15,7 @@
 @endsection
 
 @section('content')
-<h4 class="fw-bold py-3 mb-4">
+<h4 class="py-3 mb-4">
   <span class="text-muted fw-light">Extended UI /</span> Perfect
   Scrollbar
 </h4>
