@@ -3,7 +3,7 @@
 @section('title', 'Text Divider - Extended UI')
 
 @section('content')
-<h4 class="fw-bold py-3 mb-4">
+<h4 class="py-3 mb-4">
   <span class="text-muted fw-light">Extended UI /</span> Text Divider
 </h4>
 
@@ -133,4 +133,5 @@
     </div>
   </div>
   <!-- /Icons -->
-</div @endsection
+</div>
+@endsection
