@@ -3,7 +3,7 @@
 @section('title', 'Input groups - Forms')
 
 @section('content')
-<h4 class="fw-bold py-3 mb-4"><span class="text-muted fw-light">Forms /</span> Input groups</h4>
+<h4 class="py-3 mb-4"><span class="text-muted fw-light">Forms /</span> Input groups</h4>
 
 <div class="row">
   <!-- Basic -->
@@ -37,7 +37,7 @@
 
         <div class="input-group">
           <span class="input-group-text">$</span>
-          <input type="text" class="form-control" placeholder="Amount" aria-label="Amount (to the nearest dollar)" />
+          <input type="number" class="form-control" placeholder="Amount" aria-label="Amount (to the nearest dollar)" />
           <span class="input-group-text">.00</span>
         </div>
 
@@ -81,7 +81,7 @@
 
         <div class="input-group input-group-merge">
           <span class="input-group-text">$</span>
-          <input type="text" class="form-control" placeholder="100" aria-label="Amount (to the nearest dollar)" />
+          <input type="number" class="form-control" placeholder="100" aria-label="Amount (to the nearest dollar)" />
           <span class="input-group-text">.00</span>
         </div>
 
@@ -114,7 +114,6 @@
           <span class="input-group-text">@</span>
           <input type="text" class="form-control" placeholder="Username" />
         </div>
-
       </div>
     </div>
   </div>
@@ -152,14 +151,14 @@
       <h5 class="card-header">Multiple inputs & addon</h5>
 
       <div class="card-body demo-vertical-spacing demo-only-element">
-        <small class="text-light fw-semibold d-block">Multiple inputs</small>
+        <small class="text-light fw-medium d-block">Multiple inputs</small>
         <div class="input-group">
           <span class="input-group-text">First and last name</span>
           <input type="text" aria-label="First name" class="form-control">
           <input type="text" aria-label="Last name" class="form-control">
         </div>
 
-        <small class="text-light fw-semibold d-block pt-3">Multiple addons</small>
+        <small class="text-light fw-medium d-block pt-3">Multiple addons</small>
         <div class="input-group">
           <span class="input-group-text">$</span>
           <span class="input-group-text">0.00</span>
@@ -181,7 +180,7 @@
       <h5 class="card-header">Speech To Text</h5>
       <div class="card-body demo-vertical-spacing demo-only-element">
 
-        <small class="text-light fw-semibold d-block">Input Group</small>
+        <small class="text-light fw-medium d-block">Input Group</small>
 
         <div class="input-group input-group-merge speech-to-text">
           <input type="text" class="form-control" placeholder="Say it" aria-describedby="text-to-speech-addon">
@@ -190,12 +189,12 @@
           </span>
         </div>
 
-        <small class="text-light fw-semibold d-block pt-3">Textarea</small>
+        <small class="text-light fw-medium d-block pt-3">Textarea</small>
 
-        <div class="input-group input-group-merge speech-to-text">
-          <textarea class="form-control" placeholder="Say it" rows="2"></textarea>
-          <span class="input-group-text">
-            <i class='bx bx-microphone cursor-pointer text-to-speech-toggle'></i>
+        <div class="input-group input-group-merge form-send-message">
+          <textarea class="form-control message-input" placeholder="Say it" rows="2"></textarea>
+          <span class="message-actions input-group-text">
+            <i class='bx bx-microphone cursor-pointer speech-to-text'></i>
           </span>
         </div>
 
@@ -210,7 +209,7 @@
     <div class="card mb-4">
       <h5 class="card-header">Button with dropdowns & addons</h5>
       <div class="card-body demo-vertical-spacing demo-only-element">
-        <small class="text-light fw-semibold d-block">Button addons</small>
+        <small class="text-light fw-medium d-block">Button addons</small>
         <div class="input-group">
           <button class="btn btn-outline-primary" type="button" id="button-addon1">Button</button>
           <input type="text" class="form-control" placeholder="" aria-label="Example text with button addon" aria-describedby="button-addon1">
@@ -233,7 +232,7 @@
           <button class="btn btn-outline-primary" type="button">Button</button>
         </div>
 
-        <small class="text-light fw-semibold d-block pt-3">Button with dropdowns</small>
+        <small class="text-light fw-medium d-block pt-3">Button with dropdowns</small>
         <div class="input-group">
           <button class="btn btn-outline-primary dropdown-toggle" type="button" data-bs-toggle="dropdown" aria-expanded="false">Dropdown</button>
           <ul class="dropdown-menu">
