@@ -3,7 +3,7 @@
 @section('title', ' Horizontal Layouts - Forms')
 
 @section('content')
-<h4 class="fw-bold py-3 mb-4"><span class="text-muted fw-light">Forms/</span> Horizontal Layouts</h4>
+<h4 class="py-3 mb-4"><span class="text-muted fw-light">Forms/</span> Horizontal Layouts</h4>
 
 <!-- Basic Layout & Basic with Icons -->
 <div class="row">
