@@ -3,7 +3,7 @@
 @section('title', ' Vertical Layouts - Forms')
 
 @section('content')
-<h4 class="fw-bold py-3 mb-4"><span class="text-muted fw-light">Forms/</span> Vertical Layouts</h4>
+<h4 class="py-3 mb-4"><span class="text-muted fw-light">Forms/</span> Vertical Layouts</h4>
 
 <!-- Basic Layout -->
 <div class="row">
@@ -94,5 +94,4 @@
     </div>
   </div>
 </div>
-
 @endsection
