@@ -3,5 +3,5 @@
 @section('title', 'Blank layout - Layouts')
 
 @section('content')
-<h4 class="fw-bold p-4">Blank Page</h4>
+<h4 class="p-4">Blank Page</h4>
 @endsection
