@@ -3,7 +3,7 @@
 @section('title', 'Account settings - Pages')
 
 @section('content')
-<h4 class="fw-bold py-3 mb-4">
+<h4 class="py-3 mb-4">
   <span class="text-muted fw-light">Account Settings / </span> Connections
 </h4>
 
