@@ -3,7 +3,7 @@
 @section('title', 'Alerts - UI elements')
 
 @section('content')
-<h4 class="fw-bold py-3 mb-4">
+<h4 class="py-3 mb-4">
   <span class="text-muted fw-light">UI elements /</span> Alerts
 </h4>
 <div class="row mb-4">
