@@ -3,7 +3,7 @@
 @section('title', 'Collapse - UI elements')
 
 @section('content')
-<h4 class="fw-bold py-3 mb-4"><span class="text-muted fw-light">UI elements /</span> Collapse</h4>
+<h4 class="py-3 mb-4"><span class="text-muted fw-light">UI elements /</span> Collapse</h4>
 
 <!-- Collapse -->
 <h5>Collapse</h5>
