@@ -3,7 +3,7 @@
 @section('title', 'List groups - UI elements')
 
 @section('content')
-<h4 class="fw-bold py-3 mb-4">
+<h4 class="py-3 mb-4">
   <span class="text-muted fw-light">UI elements /</span> List groups
 </h4>
 
@@ -13,7 +13,7 @@
     <div class="row">
       <!-- Basic List group -->
       <div class="col-lg-6 mb-4 mb-xl-0">
-        <small class="text-light fw-semibold">Basic</small>
+        <small class="text-light fw-medium">Basic</small>
         <div class="demo-inline-spacing mt-3">
           <div class="list-group">
             <a href="javascript:void(0);" class="list-group-item list-group-item-action active">Bear claw cake biscuit</a>
@@ -27,7 +27,7 @@
       <!--/ Basic List group -->
       <!-- List group with Badges & Pills -->
       <div class="col-lg-6">
-        <small class="text-light fw-semibold">With Bagdes & Pills</small>
+        <small class="text-light fw-medium">With Bagdes & Pills</small>
         <div class="demo-inline-spacing mt-3">
           <ul class="list-group">
             <li class="list-group-item d-flex justify-content-between align-items-center">Soufflé
@@ -55,7 +55,7 @@
     <div class="row">
       <!-- List group Flush (Without main border) -->
       <div class="col-lg-6 mb-4 mb-xl-0">
-        <small class="text-light fw-semibold">Flush</small>
+        <small class="text-light fw-medium">Flush</small>
         <div class="demo-inline-spacing mt-3">
           <div class="list-group list-group-flush">
             <a href="javascript:void(0);" class="list-group-item list-group-item-action">Bear claw cake biscuit</a>
@@ -69,7 +69,7 @@
       <!--/ List group Flush (Without main border) -->
       <!-- List group Icons -->
       <div class="col-lg-6">
-        <small class="text-light fw-semibold">With Icons</small>
+        <small class="text-light fw-medium">With Icons</small>
         <div class="demo-inline-spacing mt-3">
           <ul class="list-group">
             <li class="list-group-item d-flex align-items-center">
@@ -103,7 +103,7 @@
     <div class="row">
       <!-- List group Numbered -->
       <div class="col-lg-6 mb-4 mb-xl-0">
-        <small class="text-light fw-semibold">Numbered</small>
+        <small class="text-light fw-medium">Numbered</small>
         <div class="demo-inline-spacing mt-3">
           <ol class="list-group list-group-numbered">
             <li class="list-group-item">Bear claw cake biscuit</li>
@@ -117,7 +117,7 @@
       <!--/ List group Numbered -->
       <!-- List group checkbox -->
       <div class="col-lg-6">
-        <small class="text-light fw-semibold">List Group With Checkbox</small>
+        <small class="text-light fw-medium">List Group With Checkbox</small>
         <div class="demo-inline-spacing mt-3">
           <div class="list-group">
             <label class="list-group-item">
@@ -151,7 +151,7 @@
     <div class="row">
       <!-- Contextual List group -->
       <div class="col-lg-6 mb-4 mb-xl-0">
-        <small class="text-light fw-semibold">Contextual classes</small>
+        <small class="text-light fw-medium">Contextual classes</small>
         <div class="demo-inline-spacing mt-3">
           <ul class="list-group">
             <li class="list-group-item list-group-item-primary">Primary list group item</li>
@@ -167,7 +167,7 @@
       <!--/ Contextual List group -->
       <!-- Custom content with heading -->
       <div class="col-lg-6">
-        <small class="text-light fw-semibold">Custom content</small>
+        <small class="text-light fw-medium">Custom content</small>
         <div class="demo-inline-spacing mt-3">
           <div class="list-group">
             <a href="javascript:void(0);" class="list-group-item list-group-item-action flex-column align-items-start active">
@@ -208,7 +208,7 @@
     <div class="row">
       <!-- Custom content with heading -->
       <div class="col-lg-6 mb-4 mb-xl-0">
-        <small class="text-light fw-semibold">Vertical</small>
+        <small class="text-light fw-medium">Vertical</small>
         <div class="mt-3">
           <div class="row">
             <div class="col-md-4 col-12 mb-3 mb-md-0">
@@ -255,7 +255,7 @@
         </div>
       </div>
       <div class="col-lg-6">
-        <small class="text-light fw-semibold">Horizontal</small>
+        <small class="text-light fw-medium">Horizontal</small>
         <div class="demo-inline-spacing mt-3">
           <div class="list-group list-group-horizontal-md text-md-center">
             <a class="list-group-item list-group-item-action active" id="home-list-item" data-bs-toggle="list" href="#horizontal-home">Home</a>
