@@ -3,7 +3,7 @@
 @section('title', 'Pagination and breadcrumbs - UI elements')
 
 @section('content')
-<h4 class="fw-bold py-3 mb-4">
+<h4 class="py-3 mb-4">
   <span class="text-muted fw-light">UI elements /</span> Pagination and breadcrumbs
 </h4>
 
@@ -12,8 +12,8 @@
   <!-- Basic Pagination -->
   <div class="card-body">
     <div class="row">
-      <div class="col">
-        <small class="text-light fw-semibold">Basic</small>
+      <div class="col-lg-6">
+        <small class="text-light fw-medium">Basic</small>
         <div class="demo-inline-spacing">
           <!-- Basic Pagination -->
           <nav aria-label="Page navigation">
@@ -59,7 +59,7 @@
   <div class="card-body">
     <div class="row">
       <div class="col-lg-4">
-        <small class="text-light fw-semibold">Sizes</small>
+        <small class="text-light fw-medium">Sizes</small>
         <div class="demo-inline-spacing">
           <nav aria-label="Page navigation">
             <ul class="pagination pagination-sm">
@@ -139,7 +139,7 @@
         </div>
       </div>
       <div class="col-lg-8">
-        <small class="text-light fw-semibold">Alignments</small>
+        <small class="text-light fw-medium">Alignments</small>
         <div class="demo-inline-spacing">
           <nav aria-label="Page navigation">
             <ul class="pagination">
