@@ -3,14 +3,14 @@
 @section('title', 'Progress bars - UI elements')
 
 @section('content')
-<h4 class="fw-bold py-3 mb-4"><span class="text-muted fw-light">UI elements /</span> Progress bars
+<h4 class="py-3 mb-4"><span class="text-muted fw-light">UI elements /</span> Progress bars
 </h4>
 
 <!-- Options -->
 <div class="card mb-4">
   <h5 class="card-header">Progress bars</h5>
   <div class="card-body">
-    <div class="text-light small fw-semibold">Default</div>
+    <div class="text-light small fw-medium">Default</div>
     <div class="demo-vertical-spacing">
       <div class="progress">
         <div class="progress-bar" role="progressbar" style="width: 25%;" aria-valuenow="25" aria-valuemin="0" aria-valuemax="100"></div>
@@ -22,7 +22,7 @@
   </div>
   <hr class="m-0" />
   <div class="card-body">
-    <div class="text-light small fw-semibold">Height</div>
+    <div class="text-light small fw-medium">Height</div>
     <div class="demo-vertical-spacing">
       <div class="progress" style="height: 6px;">
         <div class="progress-bar" role="progressbar" style="width: 25%;" aria-valuenow="25" aria-valuemin="0" aria-valuemax="100"></div>
@@ -34,7 +34,7 @@
   </div>
   <hr class="m-0" />
   <div class="card-body">
-    <div class="text-light small fw-semibold">With Label</div>
+    <div class="text-light small fw-medium">With Label</div>
     <div class="demo-vertical-spacing">
       <div class="progress">
         <div class="progress-bar" role="progressbar" style="width: 25%;" aria-valuenow="25" aria-valuemin="0" aria-valuemax="100">25%</div>
@@ -144,21 +144,21 @@
 <div class="card">
   <h5 class="card-header">Multiple bars</h5>
   <div class="card-body">
-    <div class="text-light small fw-semibold mb-1">Default</div>
+    <div class="text-light small fw-medium mb-1">Default</div>
     <div class="progress mb-3">
       <div class="progress-bar bg-primary shadow-none" role="progressbar" style="width: 15%" aria-valuenow="15" aria-valuemin="0" aria-valuemax="100"></div>
       <div class="progress-bar bg-success shadow-none" role="progressbar" style="width: 30%" aria-valuenow="30" aria-valuemin="0" aria-valuemax="100"></div>
       <div class="progress-bar bg-danger shadow-none" role="progressbar" style="width: 20%" aria-valuenow="20" aria-valuemin="0" aria-valuemax="100"></div>
     </div>
 
-    <div class="text-light small fw-semibold mb-1">Striped</div>
+    <div class="text-light small fw-medium mb-1">Striped</div>
     <div class="progress mb-3">
       <div class="progress-bar bg-primary progress-bar-striped shadow-none" role="progressbar" style="width: 15%" aria-valuenow="15" aria-valuemin="0" aria-valuemax="100"></div>
       <div class="progress-bar bg-success progress-bar-striped shadow-none" role="progressbar" style="width: 30%" aria-valuenow="30" aria-valuemin="0" aria-valuemax="100"></div>
       <div class="progress-bar bg-danger progress-bar-striped shadow-none" role="progressbar" style="width: 20%" aria-valuenow="20" aria-valuemin="0" aria-valuemax="100"></div>
     </div>
 
-    <div class="text-light small fw-semibold mb-1">Animated</div>
+    <div class="text-light small fw-medium mb-1">Animated</div>
     <div class="progress">
       <div class="progress-bar bg-primary progress-bar-striped progress-bar-animated shadow-none" role="progressbar" style="width: 15%" aria-valuenow="15" aria-valuemin="0" aria-valuemax="100"></div>
       <div class="progress-bar bg-success progress-bar-striped progress-bar-animated shadow-none" role="progressbar" style="width: 30%" aria-valuenow="30" aria-valuemin="0" aria-valuemax="100"></div>
