@@ -16,6 +16,8 @@
   <!-- Favicon -->
   <link rel="icon" type="image/x-icon" href="{{ asset('assets/img/favicon/favicon.ico') }}" />
 
+  
+
   <!-- Include Styles -->
   @include('layouts/sections/styles')
 
@@ -24,15 +26,13 @@
 </head>
 
 <body>
+  
+
   <!-- Layout Content -->
   @yield('layoutContent')
   <!--/ Layout Content -->
 
-  {{-- remove while creating package --}}
-  <div class="buy-now">
-    <a href="{{config('variables.productPage')}}" target="_blank" class="btn btn-danger btn-buy-now">Upgrade To Pro</a>
-  </div>
-  {{-- remove while creating package end --}}
+  
 
   <!-- Include Scripts -->
   @include('layouts/sections/scripts')
