@@ -1,80 +1,109 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
-
-/*
-|--------------------------------------------------------------------------
-| Web Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register web routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| contains the "web" middleware group. Now create something great!
-|
-*/
-
-$controller_path = 'App\Http\Controllers';
+use App\Http\Controllers\dashboard\Analytics;
+use App\Http\Controllers\layouts\WithoutMenu;
+use App\Http\Controllers\layouts\WithoutNavbar;
+use App\Http\Controllers\layouts\Fluid;
+use App\Http\Controllers\layouts\Container;
+use App\Http\Controllers\layouts\Blank;
+use App\Http\Controllers\pages\AccountSettingsAccount;
+use App\Http\Controllers\pages\AccountSettingsNotifications;
+use App\Http\Controllers\pages\AccountSettingsConnections;
+use App\Http\Controllers\pages\MiscError;
+use App\Http\Controllers\pages\MiscUnderMaintenance;
+use App\Http\Controllers\authentications\LoginBasic;
+use App\Http\Controllers\authentications\RegisterBasic;
+use App\Http\Controllers\authentications\ForgotPasswordBasic;
+use App\Http\Controllers\cards\CardBasic;
+use App\Http\Controllers\user_interface\Accordion;
+use App\Http\Controllers\user_interface\Alerts;
+use App\Http\Controllers\user_interface\Badges;
+use App\Http\Controllers\user_interface\Buttons;
+use App\Http\Controllers\user_interface\Carousel;
+use App\Http\Controllers\user_interface\Collapse;
+use App\Http\Controllers\user_interface\Dropdowns;
+use App\Http\Controllers\user_interface\Footer;
+use App\Http\Controllers\user_interface\ListGroups;
+use App\Http\Controllers\user_interface\Modals;
+use App\Http\Controllers\user_interface\Navbar;
+use App\Http\Controllers\user_interface\Offcanvas;
+use App\Http\Controllers\user_interface\PaginationBreadcrumbs;
+use App\Http\Controllers\user_interface\Progress;
+use App\Http\Controllers\user_interface\Spinners;
+use App\Http\Controllers\user_interface\TabsPills;
+use App\Http\Controllers\user_interface\Toasts;
+use App\Http\Controllers\user_interface\TooltipsPopovers;
+use App\Http\Controllers\user_interface\Typography;
+use App\Http\Controllers\extended_ui\PerfectScrollbar;
+use App\Http\Controllers\extended_ui\TextDivider;
+use App\Http\Controllers\icons\Boxicons;
+use App\Http\Controllers\form_elements\BasicInput;
+use App\Http\Controllers\form_elements\InputGroups;
+use App\Http\Controllers\form_layouts\VerticalForm;
+use App\Http\Controllers\form_layouts\HorizontalForm;
+use App\Http\Controllers\tables\Basic as TablesBasic;
 
 // Main Page Route
-Route::get('/', $controller_path . '\dashboard\Analytics@index')->name('dashboard-analytics');
+Route::get('/', [Analytics::class, 'index'])->name('dashboard-analytics');
 
 // layout
-Route::get('/layouts/without-menu', $controller_path . '\layouts\WithoutMenu@index')->name('layouts-without-menu');
-Route::get('/layouts/without-navbar', $controller_path . '\layouts\WithoutNavbar@index')->name('layouts-without-navbar');
-Route::get('/layouts/fluid', $controller_path . '\layouts\Fluid@index')->name('layouts-fluid');
-Route::get('/layouts/container', $controller_path . '\layouts\Container@index')->name('layouts-container');
-Route::get('/layouts/blank', $controller_path . '\layouts\Blank@index')->name('layouts-blank');
+Route::get('/layouts/without-menu', [WithoutMenu::class, 'index'])->name('layouts-without-menu');
+Route::get('/layouts/without-navbar', [WithoutNavbar::class, 'index'])->name('layouts-without-navbar');
+Route::get('/layouts/fluid', [Fluid::class, 'index'])->name('layouts-fluid');
+Route::get('/layouts/container', [Container::class, 'index'])->name('layouts-container');
+Route::get('/layouts/blank', [Blank::class, 'index'])->name('layouts-blank');
 
 // pages
-Route::get('/pages/account-settings-account', $controller_path . '\pages\AccountSettingsAccount@index')->name('pages-account-settings-account');
-Route::get('/pages/account-settings-notifications', $controller_path . '\pages\AccountSettingsNotifications@index')->name('pages-account-settings-notifications');
-Route::get('/pages/account-settings-connections', $controller_path . '\pages\AccountSettingsConnections@index')->name('pages-account-settings-connections');
-Route::get('/pages/misc-error', $controller_path . '\pages\MiscError@index')->name('pages-misc-error');
-Route::get('/pages/misc-under-maintenance', $controller_path . '\pages\MiscUnderMaintenance@index')->name('pages-misc-under-maintenance');
+Route::get('/pages/account-settings-account', [AccountSettingsAccount::class, 'index'])->name('pages-account-settings-account');
+Route::get('/pages/account-settings-notifications', [AccountSettingsNotifications::class, 'index'])->name('pages-account-settings-notifications');
+Route::get('/pages/account-settings-connections', [AccountSettingsConnections::class, 'index'])->name('pages-account-settings-connections');
+Route::get('/pages/misc-error', [MiscError::class, 'index'])->name('pages-misc-error');
+Route::get('/pages/misc-under-maintenance', [MiscUnderMaintenance::class, 'index'])->name('pages-misc-under-maintenance');
 
 // authentication
-Route::get('/auth/login-basic', $controller_path . '\authentications\LoginBasic@index')->name('auth-login-basic');
-Route::get('/auth/register-basic', $controller_path . '\authentications\RegisterBasic@index')->name('auth-register-basic');
-Route::get('/auth/forgot-password-basic', $controller_path . '\authentications\ForgotPasswordBasic@index')->name('auth-reset-password-basic');
+Route::get('/auth/login-basic', [LoginBasic::class, 'index'])->name('auth-login-basic');
+Route::get('/auth/register-basic', [RegisterBasic::class, 'index'])->name('auth-register-basic');
+Route::get('/auth/forgot-password-basic', [ForgotPasswordBasic::class, 'index'])->name('auth-reset-password-basic');
 
 // cards
-Route::get('/cards/basic', $controller_path . '\cards\CardBasic@index')->name('cards-basic');
+Route::get('/cards/basic', [CardBasic::class, 'index'])->name('cards-basic');
 
 // User Interface
-Route::get('/ui/accordion', $controller_path . '\user_interface\Accordion@index')->name('ui-accordion');
-Route::get('/ui/alerts', $controller_path . '\user_interface\Alerts@index')->name('ui-alerts');
-Route::get('/ui/badges', $controller_path . '\user_interface\Badges@index')->name('ui-badges');
-Route::get('/ui/buttons', $controller_path . '\user_interface\Buttons@index')->name('ui-buttons');
-Route::get('/ui/carousel', $controller_path . '\user_interface\Carousel@index')->name('ui-carousel');
-Route::get('/ui/collapse', $controller_path . '\user_interface\Collapse@index')->name('ui-collapse');
-Route::get('/ui/dropdowns', $controller_path . '\user_interface\Dropdowns@index')->name('ui-dropdowns');
-Route::get('/ui/footer', $controller_path . '\user_interface\Footer@index')->name('ui-footer');
-Route::get('/ui/list-groups', $controller_path . '\user_interface\ListGroups@index')->name('ui-list-groups');
-Route::get('/ui/modals', $controller_path . '\user_interface\Modals@index')->name('ui-modals');
-Route::get('/ui/navbar', $controller_path . '\user_interface\Navbar@index')->name('ui-navbar');
-Route::get('/ui/offcanvas', $controller_path . '\user_interface\Offcanvas@index')->name('ui-offcanvas');
-Route::get('/ui/pagination-breadcrumbs', $controller_path . '\user_interface\PaginationBreadcrumbs@index')->name('ui-pagination-breadcrumbs');
-Route::get('/ui/progress', $controller_path . '\user_interface\Progress@index')->name('ui-progress');
-Route::get('/ui/spinners', $controller_path . '\user_interface\Spinners@index')->name('ui-spinners');
-Route::get('/ui/tabs-pills', $controller_path . '\user_interface\TabsPills@index')->name('ui-tabs-pills');
-Route::get('/ui/toasts', $controller_path . '\user_interface\Toasts@index')->name('ui-toasts');
-Route::get('/ui/tooltips-popovers', $controller_path . '\user_interface\TooltipsPopovers@index')->name('ui-tooltips-popovers');
-Route::get('/ui/typography', $controller_path . '\user_interface\Typography@index')->name('ui-typography');
+Route::get('/ui/accordion', [Accordion::class, 'index'])->name('ui-accordion');
+Route::get('/ui/alerts', [Alerts::class, 'index'])->name('ui-alerts');
+Route::get('/ui/badges', [Badges::class, 'index'])->name('ui-badges');
+Route::get('/ui/buttons', [Buttons::class, 'index'])->name('ui-buttons');
+Route::get('/ui/carousel', [Carousel::class, 'index'])->name('ui-carousel');
+Route::get('/ui/collapse', [Collapse::class, 'index'])->name('ui-collapse');
+Route::get('/ui/dropdowns', [Dropdowns::class, 'index'])->name('ui-dropdowns');
+Route::get('/ui/footer', [Footer::class, 'index'])->name('ui-footer');
+Route::get('/ui/list-groups', [ListGroups::class, 'index'])->name('ui-list-groups');
+Route::get('/ui/modals', [Modals::class, 'index'])->name('ui-modals');
+Route::get('/ui/navbar', [Navbar::class, 'index'])->name('ui-navbar');
+Route::get('/ui/offcanvas', [Offcanvas::class, 'index'])->name('ui-offcanvas');
+Route::get('/ui/pagination-breadcrumbs', [PaginationBreadcrumbs::class, 'index'])->name('ui-pagination-breadcrumbs');
+Route::get('/ui/progress', [Progress::class, 'index'])->name('ui-progress');
+Route::get('/ui/spinners', [Spinners::class, 'index'])->name('ui-spinners');
+Route::get('/ui/tabs-pills', [TabsPills::class, 'index'])->name('ui-tabs-pills');
+Route::get('/ui/toasts', [Toasts::class, 'index'])->name('ui-toasts');
+Route::get('/ui/tooltips-popovers', [TooltipsPopovers::class, 'index'])->name('ui-tooltips-popovers');
+Route::get('/ui/typography', [Typography::class, 'index'])->name('ui-typography');
 
 // extended ui
-Route::get('/extended/ui-perfect-scrollbar', $controller_path . '\extended_ui\PerfectScrollbar@index')->name('extended-ui-perfect-scrollbar');
-Route::get('/extended/ui-text-divider', $controller_path . '\extended_ui\TextDivider@index')->name('extended-ui-text-divider');
+Route::get('/extended/ui-perfect-scrollbar', [PerfectScrollbar::class, 'index'])->name('extended-ui-perfect-scrollbar');
+Route::get('/extended/ui-text-divider', [TextDivider::class, 'index'])->name('extended-ui-text-divider');
 
 // icons
-Route::get('/icons/boxicons', $controller_path . '\icons\Boxicons@index')->name('icons-boxicons');
+Route::get('/icons/boxicons', [Boxicons::class, 'index'])->name('icons-boxicons');
 
 // form elements
-Route::get('/forms/basic-inputs', $controller_path . '\form_elements\BasicInput@index')->name('forms-basic-inputs');
-Route::get('/forms/input-groups', $controller_path . '\form_elements\InputGroups@index')->name('forms-input-groups');
+Route::get('/forms/basic-inputs', [BasicInput::class, 'index'])->name('forms-basic-inputs');
+Route::get('/forms/input-groups', [InputGroups::class, 'index'])->name('forms-input-groups');
 
 // form layouts
-Route::get('/form/layouts-vertical', $controller_path . '\form_layouts\VerticalForm@index')->name('form-layouts-vertical');
-Route::get('/form/layouts-horizontal', $controller_path . '\form_layouts\HorizontalForm@index')->name('form-layouts-horizontal');
+Route::get('/form/layouts-vertical', [VerticalForm::class, 'index'])->name('form-layouts-vertical');
+Route::get('/form/layouts-horizontal', [HorizontalForm::class, 'index'])->name('form-layouts-horizontal');
 
 // tables
-Route::get('/tables/basic', $controller_path . '\tables\Basic@index')->name('tables-basic');
+Route::get('/tables/basic', [TablesBasic::class, 'index'])->name('tables-basic');
