@@ -24,8 +24,9 @@ mix.options({
 mix.webpackConfig({
   output: {
     publicPath: process.env.ASSET_URL || undefined,
-    libraryTarget: 'window'
+    libraryTarget: 'umd'
   },
+
   plugins: [
     new EnvironmentPlugin({
       // Application's public url
@@ -57,7 +58,6 @@ mix.webpackConfig({
   externals: {
     jquery: 'jQuery',
     moment: 'moment',
-    'datatables.net': '$.fn.dataTable',
     jsdom: 'jsdom',
     velocity: 'Velocity',
     hammer: 'Hammer',
@@ -109,8 +109,6 @@ mixAssetsDir('vendor/libs/**/!(_)*.scss', (src, dest) =>
   mix.sass(src, dest.replace(/\.scss$/, '.css'), { sassOptions })
 );
 mixAssetsDir('vendor/libs/**/*.{png,jpg,jpeg,gif}', (src, dest) => mix.copy(src, dest));
-// Copy task for form validation plugin as premium plugin don't have npm package
-mixAssetsDir('vendor/libs/formvalidation/dist', (src, dest) => mix.copyDirectory(src, dest));
 
 // Fonts
 mixAssetsDir('vendor/fonts/*/*', (src, dest) => mix.copy(src, dest));
@@ -139,7 +137,7 @@ mix.version();
  | BrowserSync can automatically monitor your files for changes, and inject your changes into the browser without requiring a manual refresh.
  | You may enable support for this by calling the mix.browserSync() method:
  | Make Sure to run `php artisan serve` and `yarn watch` command to run Browser Sync functionality
- | Refer official documentation for more information: https://laravel.com/docs/9.x/mix#browsersync-reloading
+ | Refer official documentation for more information: https://laravel.com/docs/10.x/mix#browsersync-reloading
  */
 
 mix.browserSync('http://127.0.0.1:8000/');
